Fix picture.path backslash issue and use JwtPipe
This fixes a problem introduced by #693 that occurred on Windows
because the user.profile.picture.path then includes backslashes,
which disrupted the usage of the 'actualProfilePicturePath' as
part of the avatar URL in app.component.html. Note that it appears
that #693 lead to additional new issues / regressions that are not
fixed by this! The commit also includes some very minor refactoring
and transforms the 'actualProfilePicturePath' in app.component.ts
via the JwtPipe regarding the resolution of #729.