API: Scrap UploadsController, revert to express.static
UploadsController had issues with (at least) <video> requests
which could not be resolved so far. W.r.t. issue #729 this means
that it is currently not possible (via a @Controller) to restrict file
access for specific users to specific courses (or similar) directly.
However, since we employ randomized file names the access is actually
restricted to only those users who can obtain the secret filenames
anyway, so the goal is partially fulfilled regardless of this flaw.