API: Fix absolute picture.path issue re. #693
Note that this is supposed to be more of a temporary solution, since
the approach of using picture.path in the front-end as direct avatar
URL part (introduced by #693 i.e. PR #751) should likely be replaced
by something less error-prone in the future. (Since this has nothing
to do directly with issue #729 or PR #807, which this commit is part
of, this should get its own issue re. PR #751 aftermath.)