App: Fix AuthenticationService login code order
Calling userService.setUser before the mediaToken is set caused an
issue due to setUser currently calling updateProfilePicture, thus
causing a "chain reaction" that ultimately leads to the AppComponent
calling the jwtPipe.transform method, which fails without mediaToken.

This also removes a superfluous separate call to updateProfilePicture,
which was presumably added originally because of the previously broken
checkNewProfilePicturePath UserService method used in setUser.