Eliminate /api/duplicate/unit/:id courseId BodyParam
This parameter didn't provide any functional benefit,
assuming a lecture can only belong to a single course.
(Which can be assumed since the Units store a course id,
while being contained by lectures, thus indicating that
the lectures are also only meant to belong to one course.)
A bit of front-end code had to be adjusted for this change.