Secure routes via new INotificationSettingsView
This is primarily meant to close the vulnerabilities that the unfiltered
toObject responses caused (i.e. full data leakage of course (& user) data),
but it also changes the PutNotificationSettings route to use the @CurrentUser
instead of an arbitrary user, since INotificationSettingsView doesn't contain
the user field any longer (but this alone does not yet fully secure the request
side of that route).